Move all supertokens code into the supertokens folder
