+++ apps/web/src/app/auth/[[...path]]/page.tsx
@@ -3,7 +3,7 @@
 import { useEffect, useState } from "react";
 import { redirectToAuth } from "supertokens-auth-react";
 import SuperTokens from "supertokens-auth-react/ui";
-import { PreBuiltUIList } from "../../config/frontend";
+import { PreBuiltUIList } from "../../config/supertokens/frontend";
 
 export default function Auth() {
   // if the user visits a page that is not handled by us (like /auth/random), then we redirect them back to the auth page.
