+++ apps/web/src/app/config/supertokens/components/sessionAuthForNextJS.tsx
