+++ apps/web/src/middleware.ts
@@ -2,7 +2,7 @@ import { NextResponse } from "next/server";
 import type { NextRequest } from "next/server";
 import { SessionContainer } from "supertokens-node/recipe/session";
 import { withSession } from "supertokens-node/nextjs";
-import { ensureSuperTokensInit } from "./app/config/backend";
+import { ensureSuperTokensInit } from "./app/config/supertokens/backend";
 
 ensureSuperTokensInit();
